clustermesh: split the generic logic from the specific part
This commit refactors the clustermesh subsystem, splitting it into a
generic part (pkg/clustermesh/internal), which is responsible for watching
the configuration directory and establishing the connections to remote
etcd endpoints, and a clustermesh specific part (pkg/clustermesh),
which handles the actual synchronization of the required information).
Additionally, the helper methods dealing with the setting and retrieval
of the CiliumClusterConfig are moved to a separate `clustermesh/utils`
package. This is a preparatory change to allow reusing the generic part
for the upcoming kvstoremesh implementation, reducing code duplication and
maintenance burden.
