+++ pkg/clustermesh/internal/config.go
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright Authors of Cilium
 
-package clustermesh
+package internal
 
 import (
 	"crypto/sha256"
